ObjectListItemBase
Updated ObjectListItem so that all object list items share a common base. ObjectListItem then displays different component which extend this base, depending on the type. This makes the code more readable, and allows objects like reminders to use the same style as other objects
