+++ packages/server/src/utils/defaultPermissions.ts
@@ -4,6 +4,7 @@
  * their ModelLogic object.
  */
 export const defaultPermissions = ({ isAdmin, isDeleted, isLoggedIn, isPublic }: { isAdmin: boolean, isDeleted: boolean, isLoggedIn: boolean, isPublic: boolean }) => ({
+    canBookmark: () => isLoggedIn && !isDeleted && (isAdmin || isPublic),
     canComment: () => isLoggedIn && !isDeleted && (isAdmin || isPublic),
     canConnect: () => isLoggedIn && !isDeleted && (isAdmin || isPublic),
     canCopy: () => isLoggedIn && !isDeleted && (isAdmin || isPublic),
@@ -12,9 +13,8 @@ export const defaultPermissions = ({ isAdmin, isDeleted, isLoggedIn, isPublic }:
     canRead: () => !isDeleted && (isPublic || isAdmin),
     canReport: () => isLoggedIn && !isAdmin && !isDeleted && isPublic,
     canRun: () => !isDeleted && (isAdmin || isPublic),
-    canBookmark: () => isLoggedIn && !isDeleted && (isAdmin || isPublic),
     canTransfer: () => isLoggedIn && isAdmin && !isDeleted,
     canUpdate: () => isLoggedIn && !isDeleted && isAdmin,
     canUse: () => isLoggedIn && !isDeleted && (isAdmin || isPublic),
     canReact: () => isLoggedIn && !isDeleted && (isAdmin || isPublic),
-})
+});
