+++ packages/ui/src/components/lists/types.ts
@@ -1,4 +1,4 @@
-import { ApiVersion, CommonKey, FocusMode, GqlModelType, Meeting, NoteVersion, Organization, Project, ProjectVersion, Role, Routine, RoutineVersion, RunProject, RunRoutine, SmartContractVersion, StandardVersion, SvgComponent, Tag, User } from "@local/shared";
+import { ApiVersion, CommonKey, FocusMode, GqlModelType, Meeting, Member, NoteVersion, Organization, Project, ProjectVersion, Role, Routine, RoutineVersion, RunProject, RunRoutine, SmartContractVersion, StandardVersion, SvgComponent, Tag, User } from "@local/shared";
 import { LineGraphProps } from "components/graphs/types";
 import { AwardDisplay, NavigableObject } from "types";
 import { ObjectAction } from "utils/actions/objectActions";
@@ -15,28 +15,32 @@ export interface ObjectActionsRowProps<T extends ObjectActionsRowObject> {
     zIndex: number;
 }
 
-export interface ObjectListItemProps<T extends ListObjectType> {
+export interface ObjectListItemBaseProps<T extends ListObjectType> {
     /**
      * Callback triggered before the list item is selected (for viewing, editing, adding a comment, etc.). 
      * If the callback returns false, the list item will not be selected.
      */
     beforeNavigation?: (item: NavigableObject) => boolean | void,
+    belowSubtitle?: React.ReactNode;
+    belowTags?: React.ReactNode;
     /**
      * True if update button should be hidden
      */
     hideUpdateButton?: boolean;
-    /**
-     * Index in list
-     */
-    index: number;
     /**
      * True if data is still being fetched
      */
     loading: boolean;
     data: T | null;
     objectType: GqlModelType | `${GqlModelType}`;
+    onClick?: (dat: T) => void;
+    toTheRight?: React.ReactNode;
     zIndex: number;
 }
+export type ObjectListItemProps<T extends ListObjectType> = ObjectListItemBaseProps<T>
+export type MemberListItemProps = ObjectListItemBaseProps<Member>
+export type RunProjectListItemProps = ObjectListItemBaseProps<RunProject>
+export type RunRoutineListItemProps = ObjectListItemBaseProps<RunRoutine>
 
 export interface SortMenuProps {
     sortOptions: any[];
