+++ packages/ui/src/utils/shape/models/projectVersion.ts
@@ -27,8 +27,7 @@ export const shapeProjectVersion: ShapeModel<ProjectVersionShape, ProjectVersion
     create: (d) => ({
         ...createPrims(d, "id", "isComplete", "isPrivate", "versionLabel", "versionNotes"),
         ...createRel(d, "directories", ["Create"], "many", shapeProjectVersionDirectory),
-        ...createRel(d, "root", ["Connect", "Create"], "one", shapeProject),
-        ...createRel(d, "suggestedNextByProject", ["Connect"], "many"),
+        ...createRel(d, "root", ["Connect", "Create"], "one", shapeProject, (r) => ({ ...r, isPrivate: d.isPrivate })), ...createRel(d, "suggestedNextByProject", ["Connect"], "many"),
         ...createRel(d, "translations", ["Create"], "many", shapeProjectVersionTranslation),
     }),
     update: (o, u, a) => shapeUpdate(u, {
